Modularize home screen
